src/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yenomal"
version = "0.1.0"
description = "Shared reusable Python modules for the Yenomal monorepo."
requires-python = ">=3.10"
dependencies = [
  "einops>=0.8.0",
  "loguru>=0.7.0",
  "numpy>=1.24.0",
  "omegaconf>=2.3.0",
  "packaging>=24.0",
  "PyYAML>=6.0.0",
  "safetensors>=0.6.0",
  "timm>=1.0.0",
  "torch>=2.0.0",
  "transformers>=4.45.0",
  "wandb>=0.16.0",
]

[project.optional-dependencies]
dev = [
  "ipykernel>=6.0.0",
  "ipython>=8.0.0",
  "pytest>=8.0.0",
  "ruff>=0.12.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]

[tool.setuptools.packages.find]
where = ["."]
include = ["yenomal*"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
addopts = "-q"
